[TASK] Deprecate GeneralUtility::_GETset()
The functionality of _GETset() is really broken by design
when thinking about having a non-global state.

On top $TSFE->mergingWithGetVars(), which was mainly used
in RealURL, is deprecated, as this needs to go within a PSR-15
middlware (I actually did that in a fork already for realurl v1).

Resolves: #86389
Releases: master
Change-Id: I7081ed48131237f9d1b0da6a4ddf252a34a02072
Reviewed-on: https://review.typo3.org/58344
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>