[!!!][TASK] Don't guess extension name by looking at controller name
\TYPO3\CMS\Extbase\Mvc\Controller\AbstractController::$extensionName
had been set in the constructor by looking at the controller class
name and guessing the extension name by removing parts of the
class namespace.

Since the extension name is known when registering plugins and modules,
the extension name will no longer be guessed in the constructor but
can be fetched from the request object.

Releases: master
Resolves: #87627
Change-Id: I307275861003f7e6bd37e76a77658c283a9b2bd2
Reviewed-on: https://review.typo3.org/c/59623
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Jürgen Venne <venne@schaffrath-digital.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Jürgen Venne <venne@schaffrath-digital.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>