[BUGFIX] Do not treat all classes as entities in ClassSchema
The ClassSchema class has a property $modelType which indicates
that a class is either an entity or a value object. This property
has a default value that reflected the state of an entity.

This means, that all classes, class schemata are created for are
considered entities which then leads to the execution of further
entity related analysis code, such as looking for specific doc
blocks and/or annotations.

While this isn't an issue functionality wise, it can be in the future.
Also, the superfluous cpu cycles can and should be spared.

Releases: master, 9.5
Resolves: #87632
Change-Id: I15f7688b8bf1ac9486d77a0efc38ccb07a397b30
Reviewed-on: https://review.typo3.org/c/59806
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>