[!!!][TASK] Configure extbase request handlers via PHP
This patch removes the ability to configure extbase
request handlers via typoscript. This is done because
typoscript is too variable, i.e. the configuration may
change depending on the day, the hour and whatever
possibility typoscript offers when it comes to conditions.

The ability to configure request handlers does not vanish
but needs to be immutable and predictable at an early
stage of the runtime.

To achieve this, the configuration has to be put into the
file "EXT:Configuration/Extbase/RequestHandlers.php".

This patch is considered breaking as the configuration via
typoscript stops working immediately and the configuration
syntax slightly changed.

The easiest way to migrate to the new syntax is to have a
look at configuration files in core extensions.

With typoscript, it was also possible to override
configuration keys and therefore completely override
request handlers set by core extensions. This is no
longer possible and considered bad practice anyway.
A suitable request handler is now only determined by
its priority.

Releases: master
Resolves: #88687
Change-Id: I915d2af221f76c1ee882761f2087b985e5584bb0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61226
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>