[TASK] Deprecate Extbase's WebRequest and WebResponse
In order to simplify the request/response handling in
Extbase and to prepare the introduction of PSR-7 compatible
request and response objects, the web versions of Extbase's
request and response have been deprecated. All relevant
logic has been moved into the base classes to stay api
compatible and not introduce a breaking change.

Along with the simplification of the request and response
objects, quite a lot of instance checks have been removed
to ease the transition to having a public api that is
defined by interfaces and not concrete implementations.

Releases: master
Resolves: #89673
Change-Id: Ie920ae7d38e6de3167ff57f93fa6e98af2936d9f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62317
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>