[BUGFIX] Automatically use JsonView for json format
With https://review.typo3.org/c/Packages/TYPO3.CMS/+/59514/ all
possibilities have been taken to define which view class to use
based on the format query param. As those format query params
are still api and in use, the core must reintegrate the former
behaviour: Instantiating a JsonView object for requests with a
format query param set to 'json'.

The implementation is marked as internal as the overall idea
is to remove the format param in the long run.

Releases: master
Resolves: #90788
Change-Id: I62aaa8f44126edbe1238a5a972044f239f3362eb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63438
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>