[TASK] Avoid creation of ClassSchema instances
The extbase container provides a method isSingleton that
checks if a class implements the SingletonInterface interface.

It did so by creating a ClassSchema instance for the given
class name which is not needed as there are more efficient
ways to check for implemented interfaces that don't require
the creation of a ClassSchema instance.

Releases: master
Resolves: #90798
Change-Id: I62a3290c2d3c100a967f5a98b3d901cd61ca0db6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63820
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>