[BUGFIX] Fix several methods of ExtensionService
Due to a missing functional test and a legacy extension
configuration format in a unit test several issues with
methods in the ExtensionService class have not been
spotted.

Although the registration of plugins changed in that
manner that FQCN's are used instead of controller class
names, the internal handling in Extbase still relies on
controller names, which are internally named controller
aliases. The internal Extbase mechanics must still work
like before, passing the controller name around as this
is unique in the context of a plugin. Therefore it's not
necessary to expose or use the FQCN.

Releases: master
Resolves: #90822
Change-Id: Idd574bcbcddaa209337f73ad8526b2423ad9c786
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63933
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Toben Schmidt <t4ss5t@posteo.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Toben Schmidt <t4ss5t@posteo.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>