[TASK] Clean up ExtensionService in Extbase
ExtensionService (internal API) is a random place for some methods used
only in one specific parts of Extbase.

Some parts in Extbase internal code regarding request building
and determining if an action is cacheable is moved to the places
where it is needed.

The EnvironmentService class has a method to determine a
HTTP-relevant request, which is only relevent for the RequestBuilder
where it is now moved to, and uses PSR-7 handling if available.

Resolves: #91019
Releases: master
Change-Id: Ie6ca0c751bc7453de1d123665b7fe6ce51b37c62
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64157
Tested-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>