[BUGFIX] Fix detection of plugin name by action
While switching from controller class aliases to fully
qualified controller class names (internally), the structure
of framework configuration slightly changed.

This unfortunately broke the detection of the plugin name
by action because the new structure had not been checked for.

Also, there was a unit test which hasn't been adjusted to the
new structure as well because it was simply overlooked.

With this patch the bug is fixed and the fix is also covered
again by a unit test.

Releases: master
Resolves: #91249
Change-Id: I9acdaeb66010563d82c818ddf0a73e2adab780db
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64388
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>