[BUGFIX] Do not convert native DATETIME values in Extbase
In contrast to the actual documentation in the code, when using
RDBMS native DATE / TIME / DATETIME fields, the RDBMS does not
adjust these field's values for time zone.
Independent of the actual time zone the DMBS (or the current
session) is using, the data is stored and retrieved "as is".

The only thing that matters is the TZ of PHP, which must be
identical for storing and retrieving the data, otherwise
data is modified.

TYPO3 works just fine like that in Backend, however Extbase'
DataMapper considered the timezone of these fields as "UTC-stored",
which is wrong.

This patch corrects this behaviour to match the rest of the Core.

Resolves: #91240
Releases: master
Change-Id: Idb8e0a6f241aa0ed4e8d4a66129b58e34b9b3292
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64053
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Thorben Kapp <thorben@webcoast.dk>
Tested-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Thorben Kapp <thorben@webcoast.dk>
Reviewed-by: Alexander Schnitzler <git@alexanderschnitzler.de>
Reviewed-by: Benni Mack <benni@typo3.org>