[BUGFIX] Ensure proper encapsulation of comparisons with QueryBuilder
When using a TCA table without deleted, or any other enablefields
except "endtime", then the SQL query is built in a wrong fashion
both in PageRepository and BackendUtility. Thee source of this disaster
is inproper encapsulation in parenthesis when comparisons are build.

Instead of simply adding a manual "()" around the or() comparisons,
this change adjusts CompositionExpresion to properly add corresponding
parenthesis, which ensure this in every case and avoid missing usages.

A couple of tests are adjusted to properly reflect this change.

Resolves: #89616
Releases: main
Change-Id: I4aed140cbca5b4b40dfacd459a4c3aa30549b582
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76481
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>