[TASK] Make Argument and Arguments classes dependency free
Classes Argument and Arguments are prototypes (stateful) which
have been created by the object manager to have dependencies
injected. This however is problematic for various reasons:

1) ObjectManager shouldn't be injected into classes as it's
   used to fetch objects (with DI support) directly from the
   container.
2) In those classes, prototypes had been injected which are now
   created directly via makeInstance.

To be able to achieve this, the property mapper had been pulled
out of the Argument class and does not longer manipulate the
internal value on a setValue call. The proper value is now set
from outside the Argument instances in the ActionController.

Releases: master
Resolves: #92236
Change-Id: Ic0ca6b84c003c6d15c9fde4fa64cb3bbb71f7880
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65640
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>