[!!!][TASK] Remove fluid widgets
All fluid widgets and the core widget functionality has
been removed due to its complexity and the fact that
fluid widgets did and do bring more trouble than benefit.

The most commonly used paginate widget can be replaced with
a recently introduced new core api. Examples of how to use
it can be found in the core itself. Extensions be_user and
extensionmanager did use the pagination widget before and
do now use the core pagination api.

Custom widgets and their functionality have to be
reimplemented in another way. The usual way to replace a
widget is to put its functionality into the controller
action and move the widgets query params into the query
param namespace of the plugin that formerly rendered the
widget.

Releases: master
Resolves: #92529
Change-Id: I3342214d467a85dfcd6d684f4a7678f70fd67146
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/66095
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>