[BUGFIX] Silently resolve invalid current page in paginator
Passing a current page number which exceeds the maximum
number of pages to the paginator is now silently resolved
by using the last available page as the current page. This
is done to prevent empty results, since such value can be
passed unintentionally by a user. For example in case some
internal state changed after the view was initially rendered
or the view does not perform checks on this value at all.

The later will now be prevented at least in core by adding
the "max" attribute to all core paginations which previously
missed it.

Resolves: #93351
Releases: master, 10.4
Change-Id: I56fa1bd0ea62b7ddf7ea049819efc403fdf70d10
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67712
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>