[TASK] Use native PSR-7 Request in Extbase Request building
Because the original PSR-7 request is now
handed into Extbase at any times, the RequestBuilder
can now utilize a PSR-7 request and does not need
to fall back to global GET/POST or getIndpEnv()
methods anymore.

The final logic in this part is to handle $_FILES,
which needs to be dealt with separately as a feature
toggle.

Resolves: #93853
Releases: master
Change-Id: I60d592e6adcd705397b0f753224ee53fdf16cc0c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67519
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>