[TASK] Deprecate extbase request getBaseUri()
To further prepare extbase towards a extbase Request
object compatible with PSR-7 ServerRequestInterface, we
need some solution for those methods within existing
extbase Request that prevent immutability.

One relatively straight case are setBaseUri() and
getBaseUri(): The setter has been marked @internal in
v10 already and can be dropped, and the getter is used
just in a couple of cases.

The patch adapts usages of getBaseUri() to implement
the uri retrieval on their own based on the PSR-7
request (global access as temporary solution) and
deprecates the method. The setter is dropped.

Change-Id: Ifc224ecb8ab1a9d2b6f789fe8b92b692187614b7
Resolves: #94223
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69310
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>