[TASK] Improve exception messages in ImageService
The Extbase ImageService contains a helper method
`getImage()`. This method tries to find an image in a
couple of different ways, depending on the given input
arguments.

To improve the usability, the exception messages are
now dedicated to their corresponding try. As a side
effect, the method is now also more readable as it
now uses guard clauses and also contains a couple
of code comments.

Resolves: #94518
Releases: master, 10.4
Change-Id: I31d94eb89dd58a01d7911907ca82a256a0ba7cf3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69792
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>