[TASK] Move cache clearing into CacheService of Extbase
Currently Extbase ships with a custom CacheService
which contains all PageIDs where the cache should be
cleared once the plugin has been finished processing.

In order to reduce the amount of injections within
Typo3DbBackend, and decoupled Typo3DbBackend from
the actual implementation of how caches are flushed,
the cache clearing for pages of specific records
is now moved into CacheService directly.

This way, a better separation of concerns happens
and custom StorageBackends can be implemented without
needing to worry about TYPO3's internal cache flushing
strategy.

On top, it will become simpler to remove the SingletonInterface
of Typo3DbBackend, if desired, as it does not keep
state of the plugin via ConfigurationManager.

Additionally, if we ever want to create a generic
Cache Clearing logic (decoupled from DataHandler),
we could simply combine CacheService with the new API.

Resolves: #94703
Releases: master
Change-Id: Ia79c01c35e6440634d2018f890f661acabacd852
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70186
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>