[TASK] Unify core cache identifiers
Core cache is used in many places throughout the core,
mostly to cache resources gathered from installed
extensions. In every single place the cache identifier
for these caches is constructed in a similar, but
not identical way.

This change aims to unify the cache identifiers and
introduce a (private) API to generate a cache identifier
same way for all core caches.

As a "side effect" all identifier suffixes are now
identical and can be recognized to belong together.

Releases: master
Resolves: #95024
Change-Id: I0e090af636c58795473059191f1e571d5a6c1b73
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70810
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>