[BUGFIX] Keep extbase attributes in OriginalRequest
When an Extbase action returns a ForwardResponse, the
extbase attributes of the originalRequest are overwritten,
because the resulting request object is cloned from the
current request which holds a reference to extbase
attributes.

This change ensures, that the extbase attributes
remain in their original state in the originalRequest
object.

Resolves #95095
Releases: master
Change-Id: I25d9c3451146ec06e13bc207e52dc76024892fd0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71133
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>