[BUGFIX] Extbase UriBuilder warning in Backend context
The extbase UriBuilder can be used in backend context,
ActionMenuItemViewHelper does that for instance. The
code accesses $GLOBALS['TSFE'] at one point, which
may not be set.

The ActionMenuItemViewHelperTest triggers this, but
it should be a general asset of UriBuilder to work
without TSFE, so this is not considered an incomplete
test setup, but fixed in UriBuilder directly.

Resolves: #95529
Releases: master
Change-Id: I59bcdf00b11fa0f1af13207021da8d7ef3e4c0f4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71491
Tested-by: core-ci <typo3@b13.com>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>