Revert "[TASK] Deprecation free Serializable usage in ObjectStorage"
This reverts commit c7c47956c054ef68e2847b06f15cf69b4025bc2c.

The patch did not make sense: A class that has methods
serialize() and unserialize() and NOT implements \Serializable
interface, will never get these methods invoked when
php function serialize() or unserialize() is called.

So, the two methods have been bogus ever since.

By renaming serialize() to __serialize() and unserialize()
to __unserialize(), we actively activated the exception.
This is reverted now. And both method now have a @todo to
drop them in v12.

Reverts: #95712
Releases: master
Change-Id: I948d4b057a00ac9d53b30de55792295c1e7ae02a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71904
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>