[BUGFIX] Deprecation free \Iterator in LazyLoadingProxy
PHP 8.1 added concrete return types to internal
interfaces, classes and methods which triggers
deprecation messages, or fatal if it is related
to inheritance or interface implementations.

There are two ways to mitigate this:

* add concrete return type
* add #[\ReturnTypeWillChange]

Adding concrete return types would be breaking,
thus this patch adds #[\ReturnTypeWillChange]
to methods related to the \Iterator interface,
which LazyLoadingProxy implements.

Further @todo comments are added to remove
the php attribute #[\ReturnTypeWillChange]
and add concrete return types as breaking
change in v12.

Resolves: #95750
Releases: master
Change-Id: Ibd18158928c6290c6a04e50e073c1ec6a58a980d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71953
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Gilli <typo3@gilbertsoft.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>