[TASK] Clean up unrealistic @todo in QueryInterface
There are two QueryInterface methods marked with @todo
"See if we can get rid of these" for a very long time.
It doesn't look feasible to get rid of QuerySettings
any time soon, the construct will survive for the time
being and may only be tackled when we have a broader
look at the extbase query logic. Remove the todos.

Resolves: #96295
Releases: main
Change-Id: I5b57774b17ddebc0ecb6754b43e239880f631e4f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72562
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>