[TASK] Avoid importDataSet() in functional tests
This avoids using importDataSet() in functional tests
using .xml fixture files from the testing framework.

Usages are converted to importCSVDataSet(), fixture
files are added as .csv to the according extension.

XML handling in functional tests is now down to
setUpBackendUserFromFixture() which will be prepared
in testing-framework and adapted in core with another
patch.

Change-Id: Ifc3cc30d4b824775dc6ccc7e1d982d9d122f39d9
Resolves: #97924
Related: #96856
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75165
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>