[BUGFIX] Correct bad doc type in QueryObjectModelFactory
The docblock for QueryObjectModelFactory::comparison() expects
a string, specifically one of the QueryObjectModelConstants.JCR_OPERATOR_*
constants.  However, that class seems to no longer exist,
and what it actually uses is an integer. (Passing a string through
would break in subsequent sub-calls.)

This corrects the type to the correct integer constants that
are actually used in practice, and uses a language type
to ensure it is used properly.

Used command:

> ./Build/Scripts/runTests.sh -s phpstanGenerateBaseline

Resolves: #98007
Releases: main, 11.5
Change-Id: Id00e30370a1944a3158cb46825ccc4e15174f7ab
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75303
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>