[!!!][TASK] Remove uses of the Serializable interface
The Serializable interface has been deprecated, and is
slated for removal entirely in PHP 9. The preferred
serialization tool is the __serialize/__unserialize
method pair, which is already implemented and used
instead of Serializable. This patch removes the
already-deprecated and vestigial Serializable interface.

Serializable and __serialize() use a different string format,
so the fixture file needed to be updated as well.

Resolves: #98032
Releases: main
Change-Id: Icd021d7d48252d2955491a97201547f42e4d20e9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75306
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>