[BUGFIX] Do not query database for non-existent records
Extbase does not ask the database anymore for relations, where the uid
in the database field is zero.

This prevents SQL queries with where-parts like `uid IN (0)`.

Since this is done when reconstituting models from the database,
depending on the amount of relations of a model, the number of saved
queries is huge.

Resolves: #99505
Releases: main, 11.5
Change-Id: If3b6c2756674eeb43a66b59a13e44011c9845cd5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77582
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stephan Großberndt <stephan.grossberndt@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Stephan Großberndt <stephan.grossberndt@typo3.org>