[TASK] Use local Request Object in Extbase UriBuilder
The EXT:extbase UriBuilder already has a local instance
of the RequestInterface object, but relies in its method
`buildBackendUrl()` solely on $GLOBALS['TYPO3_REQUEST'].

This is now cleaned up, with a fallback on the global object
in case none has been given as a local instance. In this case,
a deprecation warning will be raised.

Resolves: #99650
Releases: main
Change-Id: Ib05a92c999e83e19cfde65a038439e3f571afba7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77507
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>